feat: Filter - Sort - Search.

+++ pages/api/categories/[id].js
@@ -1,5 +1,6 @@
 import connectDB from '../../../utils/connectDB';
 import Categories from '../../../models/modelCategories';
+import Products from '../../../models/modelProduct';
 import auth from '../../../middleware/auth';
 
 connectDB();
@@ -48,6 +49,12 @@ const deleteCategory = async (req, res) => {
     }
 
     const { id } = req.query;
+
+    const products = await Products.findOne({ category: id });
+
+    if (products) {
+      return res.status(400).json({ err: 'Please delete all products with a relationship.' });
+    }
 
     await Categories.findByIdAndDelete(id);
